Replace `PreviewType` with an `int`
The value for preview is the actual minute the preview is in.
The preview isn't updated to the client, other than when it's being set to `-1` (off) and `0`.
But from the app, when the user moves the slider to a timestamp manually, the preview will be sent to MQTT, in that case another value than `-1`/`0`